Add Rules class
