+++ spec/game_spec.rb
@@ -65,57 +65,4 @@ describe Game do
       expect(console.print_message(console.printed_string)).to eq(expected_output)
     end
   end
-
-  describe 'game_over?' do
-    it 'returns true if there is a winning board' do
-      board = Board.new
-      console = FakeConsole.new
-      game = Game.new(board, console)
-
-      board.mark(6)
-      board.mark(7)
-      board.mark(8)
-
-      game.game_over?
-      expect(game.game_over?).to eq(true)
-    end
-  end
-
-  describe 'game_over?' do
-    it 'returns true if there is a full board' do
-      board = Board.new
-      console = FakeConsole.new
-      game = Game.new(board, console)
-
-      board.mark(0)
-      board.mark(1)
-      board.mark(2)
-      board.mark(3)
-      board.mark(4)
-      board.mark(5)
-      board.mark(6)
-      board.mark(7)
-      board.mark(8)
-
-      game.game_over?
-      expect(game.game_over?).to eq(true)
-    end
-  end
-
-  describe 'game_over?' do
-    it 'returns false if there is not a full board or a winning board' do
-      board = Board.new
-      # console = FakeConsole.new
-      game = Game.new(board)
-
-      board.mark(0)
-      board.mark(1, 'O')
-      board.mark(2)
-      board.mark(7, 'O')
-      board.mark(8)
-
-      game.game_over?
-      expect(game.game_over?).to eq(false)
-    end
-  end
 end
